add a lot functions

+++ src/app/pages/cartes/cartes.component.ts
@@ -7,10 +7,17 @@ import { CartesService } from './cartes.service';
 })
 export class CartesComponent implements OnInit {
   cartes;
+  mc:string = "" ;
+  page:number = 0;
+  size:number = 5;
   constructor(private service: CartesService) { }
 
   ngOnInit() {
-    this.service.getCartes()
+    this.onGetAllCartes();
+    this.chercher();
+  }
+  onGetAllCartes(){
+    this.service.getCartes(this.mc,this.size,this.page)
     .subscribe(data => {
       this.cartes = data;
     }, err => {
@@ -18,12 +25,18 @@ export class CartesComponent implements OnInit {
     });
   }
   deleteCarte(id) {
+    let c=confirm("Etes vous sur?");
+    if(!c) return;
     this.service.deleteCarte(id)
     .subscribe(data => {
-
+      this.onGetAllCartes();
     }, err => {
       console.log(err);
     });
+  }
+
+  chercher(){
+    this.onGetAllCartes();
   }
 
 }
